requirements.txt
# Root dependencies (used by weekly-sync GitHub Action for article ingestion)
# Includes everything in api/requirements.txt plus ingestion extras

# API core (keep in sync with api/requirements.txt)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
anthropic>=0.18.0
openai>=1.12.0
supabase>=2.5.0
python-dotenv>=1.0.0
pydantic>=2.6.0

# Officials import (scripts/import_officials.py)
pyyaml>=6.0
requests>=2.31.0

# Article ingestion (scripts/ingest_articles.py)
beautifulsoup4>=4.12.0
lxml>=5.1.0
httpx>=0.27.0
python-dateutil>=2.8.0
